# 04/transport/api/openapi.yaml
openapi: 3.0.3
info:
  title: User management service
  description: User management service allowing to manage users.
  version: 0.0.0
servers:
  - url: http://localhost
    description: Localhost server running on HTTP.

paths:
  "/ping":
    get:
      description: Health check
      operationId: Ping
      summary: Check the healthiness of the service
      responses:
        "204":
          description: It's ALIVE (a.k.a. service is up and running)
        "400":
          description: Invalid request
        "500":
          description: Internal server error
  "/users":
    get:
      description: List all users
      operationId: ListUsers
      summary: Get all users
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "500":
          description: Internal server error
    post:
      description: Create a new user
      operationId: CreateUser
      summary: Create user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        "201":
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid request
        "500":
          description: Internal server error
  "/users/{email}":
    parameters:
      - name: email
        in: path
        required: true
        schema:
          type: string
          format: email
        description: Email of the user
    get:
      description: Get user by email
      operationId: GetUser
      summary: Get user
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: User not found
        "500":
          description: Internal server error
    put:
      description: Update user by email
      operationId: UpdateUser
      summary: Update user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        "200":
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid request
        "404":
          description: User not found
        "500":
          description: Internal server error
    delete:
      description: Delete user by email
      operationId: DeleteUser
      summary: Delete user
      responses:
        "204":
          description: User deleted successfully
        "404":
          description: User not found
        "500":
          description: Internal server error

components:
  schemas:
    User:
      type: object
      required:
        - email
        - firstName
        - lastName
      properties:
        email:
          type: string
          format: email
          description: Email address of the user
        firstName:
          type: string
          description: First name of the user
        lastName:
          type: string
          description: Last name of the user